Allow passing a 'timeout' parameter to the client constructor.
This will be passed to the Session.send() call, allowing a timeout to be
set once for a client, and affect all requests.

It is also possible to pass in a timeout value to an individual call,
which will be used just for that call. Passing in None will result in
no timeout.

This also drops support for older python versions: travis is unable to
manage installing the correct dependencies for those.